# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: SAFe Tooling API
  version: "1.0.0"
  description: |
    REST API for the SAFe Tooling platform. Manages ARTs, Teams, Program
    Increments, Features, Stories, PI Objectives, Risks, Dependencies, and
    Capacity Plans.

    All entity IDs are UUID strings. Dates are ISO 8601 strings (YYYY-MM-DD).

    Computed fields (wsjf_score, cost_of_delay, available_capacity) are
    returned in responses but must not be sent in request bodies.
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /iterations:
    get:
      tags: [iterations]
      summary: List iterations
      operationId: list_iterations
      parameters:
        - name: pi_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: PI to list iterations for
      responses:
        "200":
          description: Iterations sorted by number
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Iteration"

    post:
      tags: [iterations]
      summary: Add an iteration to a PI
      operationId: create_iteration
      description: |
        Iteration dates must fall within the PI's date range.
        Also appends the iteration id to PI.iteration_ids.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IterationCreate"
      responses:
        "201":
          description: Created iteration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Iteration"
        "404":
          description: PI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
  /iterations/{iteration_id}:
    parameters:
      - $ref: "#/components/parameters/iteration_id"
    get:
      tags: [iterations]
      summary: Get an iteration
      operationId: get_iteration
      responses:
        "200":
          description: Iteration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Iteration"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [iterations]
      summary: Update an iteration
      operationId: update_iteration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IterationUpdate"
      responses:
        "200":
          description: Updated iteration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Iteration"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [iterations]
      summary: Delete an iteration
      operationId: delete_iteration
      description: Also removes the iteration id from PI.iteration_ids.
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Features
  # ---------------------------------------------------------------------------
  /features:
    get:
      tags: [features]
      summary: List features
      operationId: list_features
      parameters:
        - name: pi_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by PI
        - name: team_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by team
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/FeatureStatus"
          description: Filter by status
        - name: sort
          in: query
          schema:
            type: string
            enum: [wsjf_desc, name_asc]
            default: wsjf_desc
          description: "wsjf_desc = ranked by WSJF score (default), name_asc = alphabetical"
      responses:
        "200":
          description: List of features
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feature"

    post:
      tags: [features]
      summary: Add a feature to the backlog
      operationId: create_feature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureCreate"
      responses:
        "201":
          description: Created feature (includes computed cost_of_delay and wsjf_score)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "422":
          $ref: "#/components/responses/ValidationError"
  /features/{feature_id}:
    parameters:
      - $ref: "#/components/parameters/feature_id"
    get:
      tags: [features]
      summary: Get a feature
      operationId: get_feature
      responses:
        "200":
          description: Feature details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [features]
      summary: Update a feature
      operationId: update_feature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureUpdate"
      responses:
        "200":
          description: Updated feature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [features]
      summary: Delete a feature
      operationId: delete_feature
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Stories
  # ---------------------------------------------------------------------------
...
  /objectives:
    get:
      tags: [objectives]
      summary: List PI objectives
      operationId: list_objectives
      parameters:
        - name: pi_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by PI
        - name: team_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by team
        - name: is_stretch
          in: query
          schema:
            type: boolean
          description: "true = stretch only, false = committed only"
      responses:
        "200":
          description: List of PI objectives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PIObjective"

    post:
      tags: [objectives]
      summary: Create a PI objective
      operationId: create_objective
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PIObjectiveCreate"
      responses:
        "201":
          description: Created PI objective
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PIObjective"
        "422":
          $ref: "#/components/responses/ValidationError"
  /objectives/{objective_id}:
    parameters:
      - $ref: "#/components/parameters/objective_id"
    get:
      tags: [objectives]
      summary: Get a PI objective
      operationId: get_objective
      responses:
        "200":
          description: PI objective details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PIObjective"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [objectives]
      summary: Update a PI objective (includes scoring actual business value)
      operationId: update_objective
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PIObjectiveUpdate"
      responses:
        "200":
          description: Updated PI objective
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PIObjective"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [objectives]
      summary: Delete a PI objective
      operationId: delete_objective
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Risks
  # ---------------------------------------------------------------------------
  /risks:
    get:
      tags: [risks]
      summary: List risks
      operationId: list_risks
      parameters:
        - name: pi_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by PI
        - name: team_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by team
        - name: roam_status
          in: query
          schema:
            $ref: "#/components/schemas/ROAMStatus"
          description: Filter by ROAM status
      responses:
        "200":
          description: List of risks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Risk"

    post:
      tags: [risks]
      summary: Raise a risk
      operationId: create_risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskCreate"
      responses:
        "201":
          description: Created risk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "422":
          $ref: "#/components/responses/ValidationError"
  /risks/{risk_id}:
    parameters:
      - $ref: "#/components/parameters/risk_id"
    get:
      tags: [risks]
      summary: Get a risk
      operationId: get_risk
      responses:
        "200":
          description: Risk details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [risks]
      summary: Update a risk (includes ROAM disposition)
      operationId: update_risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskUpdate"
      responses:
        "200":
          description: Updated risk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [risks]
      summary: Delete a risk
      operationId: delete_risk
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Dependencies
  # ---------------------------------------------------------------------------
...
components:
...
  parameters:
...
    iteration_id:
      name: iteration_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    feature_id:
      name: feature_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    objective_id:
      name: objective_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    risk_id:
      name: risk_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "PI 'abc-123' not found"

    # --- Enums ---
...
    FeatureStatus:
      type: string
      enum: [funnel, analyzing, backlog, implementing, done]
...
    ROAMStatus:
      type: string
      enum: [unroamed, resolved, owned, accepted, mitigated]
...
    Iteration:
      type: object
      required: [id, pi_id, number, start_date, end_date, is_ip]
      properties:
        id:
          type: string
          format: uuid
        pi_id:
          type: string
          format: uuid
        number:
          type: integer
          minimum: 1
        name:
          type: string
          default: ""
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        is_ip:
          type: boolean
          default: false
          description: True for the Innovation & Planning iteration
    IterationCreate:
      type: object
      required: [pi_id, number, start_date, end_date]
      properties:
        pi_id:
          type: string
          format: uuid
        number:
          type: integer
          minimum: 1
        name:
          type: string
          default: ""
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        is_ip:
          type: boolean
          default: false
    IterationUpdate:
      type: object
      properties:
        number:
          type: integer
          minimum: 1
        name:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        is_ip:
          type: boolean

    # --- Feature ---
    Feature:
      type: object
      required:
        - id
        - name
        - status
        - user_business_value
        - time_criticality
        - risk_reduction_opportunity_enablement
        - job_size
        - cost_of_delay
        - wsjf_score
        - story_ids
        - dependency_ids
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Single Sign-On
        description:
          type: string
          default: ""
        pi_id:
          type: string
          format: uuid
          nullable: true
        team_id:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/FeatureStatus"
        acceptance_criteria:
          type: string
          default: ""
        user_business_value:
          type: integer
          minimum: 1
          maximum: 10
        time_criticality:
          type: integer
          minimum: 1
          maximum: 10
        risk_reduction_opportunity_enablement:
          type: integer
          minimum: 1
          maximum: 10
        job_size:
          type: integer
          minimum: 1
          maximum: 13
        cost_of_delay:
          type: integer
          readOnly: true
          description: "Computed: user_business_value + time_criticality + risk_reduction_opportunity_enablement"
        wsjf_score:
          type: number
          format: float
          readOnly: true
          description: "Computed: cost_of_delay / job_size (rounded to 2 dp)"
        story_ids:
          type: array
          items:
            type: string
            format: uuid
        dependency_ids:
          type: array
          items:
            type: string
            format: uuid
    FeatureCreate:
      type: object
      required:
        - name
        - user_business_value
        - time_criticality
        - risk_reduction_opportunity_enablement
        - job_size
      properties:
        name:
          type: string
        description:
          type: string
          default: ""
        pi_id:
          type: string
          format: uuid
          nullable: true
        team_id:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/FeatureStatus"
          default: backlog
        acceptance_criteria:
          type: string
          default: ""
        user_business_value:
          type: integer
          minimum: 1
          maximum: 10
        time_criticality:
          type: integer
          minimum: 1
          maximum: 10
        risk_reduction_opportunity_enablement:
          type: integer
          minimum: 1
          maximum: 10
        job_size:
          type: integer
          minimum: 1
          maximum: 13
    FeatureUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        pi_id:
          type: string
          format: uuid
          nullable: true
        team_id:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/FeatureStatus"
        acceptance_criteria:
          type: string
        user_business_value:
          type: integer
          minimum: 1
          maximum: 10
        time_criticality:
          type: integer
          minimum: 1
          maximum: 10
        risk_reduction_opportunity_enablement:
          type: integer
          minimum: 1
          maximum: 10
        job_size:
          type: integer
          minimum: 1
          maximum: 13

    # --- Story ---
...
    PIObjective:
      type: object
      required:
        - id
        - description
        - team_id
        - pi_id
        - planned_business_value
        - is_stretch
        - feature_ids
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          example: Ship the authentication feature
        team_id:
          type: string
          format: uuid
        pi_id:
          type: string
          format: uuid
        planned_business_value:
          type: integer
          minimum: 1
          maximum: 10
        actual_business_value:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
        is_stretch:
          type: boolean
          default: false
          description: "Committed when false, stretch when true"
        feature_ids:
          type: array
          items:
            type: string
            format: uuid
    PIObjectiveCreate:
      type: object
      required: [description, team_id, pi_id, planned_business_value]
      properties:
        description:
          type: string
        team_id:
          type: string
          format: uuid
        pi_id:
          type: string
          format: uuid
        planned_business_value:
          type: integer
          minimum: 1
          maximum: 10
        actual_business_value:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
        is_stretch:
          type: boolean
          default: false
        feature_ids:
          type: array
          items:
            type: string
            format: uuid
          default: []
    PIObjectiveUpdate:
      type: object
      properties:
        description:
          type: string
        planned_business_value:
          type: integer
          minimum: 1
          maximum: 10
        actual_business_value:
          type: integer
          minimum: 1
          maximum: 10
          nullable: true
        is_stretch:
          type: boolean
        feature_ids:
          type: array
          items:
            type: string
            format: uuid

    # --- Risk ---
    Risk:
      type: object
      required: [id, description, pi_id, roam_status, raised_date]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          example: Auth service may not be ready in time
        pi_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
          nullable: true
        feature_id:
          type: string
          format: uuid
          nullable: true
        roam_status:
          $ref: "#/components/schemas/ROAMStatus"
        owner:
          type: string
          nullable: true
          example: Jane Smith
        mitigation_notes:
          type: string
          default: ""
        raised_date:
          type: string
          format: date
    RiskCreate:
      type: object
      required: [description, pi_id]
      properties:
        description:
          type: string
        pi_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
          nullable: true
        feature_id:
          type: string
          format: uuid
          nullable: true
        roam_status:
          $ref: "#/components/schemas/ROAMStatus"
          default: unroamed
        owner:
          type: string
          nullable: true
        mitigation_notes:
          type: string
          default: ""
    RiskUpdate:
      type: object
      properties:
        description:
          type: string
        team_id:
          type: string
          format: uuid
          nullable: true
        feature_id:
          type: string
          format: uuid
          nullable: true
        roam_status:
          $ref: "#/components/schemas/ROAMStatus"
        owner:
          type: string
          nullable: true
        mitigation_notes:
          type: string

    # --- Dependency ---
...